refactor: split main.ts into rankings-view.ts and city-detail-view.ts
Closes #158
