+++ CLAUDE.md
@@ -235,7 +235,9 @@ Primary data source is `game-defs.json` generated by the parser. Manual data ent
 
 ```
 /src/frontend       - TypeScript source (compiled by Vite)
-  main.ts           - rankings page: city rankings, city detail with fleet recommendations
+  main.ts           - rankings page orchestrator: init, routing, shared state
+  rankings-view.ts  - rankings table rendering, search/filter, country dropdown, score tiers
+  city-detail-view.ts - city detail rendering, fleet display, export buttons, garage toggle
   data.ts           - barrel re-export from sub-modules (backward compatibility)
   types.ts          - TypeScript interfaces (City, Company, Cargo, Trailer, etc.)
   loader.ts         - loadAllData(), entity builders, JSON fetching
